fix:Setmeal 模块中，批量删除套餐(delete 方法)修复(对 validIds.size() 进行判断)

+++ sky-take-out/sky-server/src/main/java/com/sky/service/impl/SetmealServiceImpl.java
@@ -208,7 +208,9 @@ public class SetmealServiceImpl implements SetmealService {
             }
         }
         // Mapper 层，批量删除
-        setmealMapper.deleteByIds(validIds);
+        if (validIds.size() > 0) {
+            setmealMapper.deleteByIds(validIds);
+        }
 
         if (!"".equals(msg.toString())) {
             // 部分启售中的套餐不能删除
